Fix bugs in merge_randoop_xml.py

+++ projects/3_gaj/evodoop.sh
@@ -1,3 +1,8 @@
+if [[ ! $# -eq 1 ]] ; then
+    echo 'Please input the number of test suites you want to generate for each of the tools!'
+    exit 1
+fi
+
 find ./src/main/java -type f > all_class_names_temp.txt
 
 rm -f all_class_names.txt
@@ -12,9 +17,10 @@ done < all_class_names_temp.txt
 
 rm -rf ./test_suites
 
-for i in {1..10}
+for i in {1..$1}
 do
   ./randoop_generation_once.sh $i
+  python3 ./merge_randoop_xml.py $i
   ./evosuite_generation_once.sh $i
 done
 rm -f all_class_names_temp.txt
